Set the document title in the PDF viewer to be the name of the PDF file.
When printing to a PDF from a PDF document, the title of the page gets used.
Currently this defaults to the chrome-extension:// URL which is confusing.
Instead we use the actual name of the PDF as the title of the extension page
which gives the appropriate filename when saving the PDF file from the print
menu.

BUG=435478

Review URL: https://codereview.chromium.org/814763002

Cr-Commit-Position: refs/heads/master@{#309108}