OOP PDF: Use the default browser zoom level.
Previously, the PDF extension assumed a default browser zoom level of
100%. https://crrev.com/324570 changed the extension to not call
chrome.tabs.setZoom() when the browser zoom level was already set to the
extension's zoom level. This caused a mismatch when the default browser
zoom was not set to 100% and the extension chose 100% as the zoom level
for the PDF. This CL changes the extension to use the default browser
zoom level.

BUG=478038

Review URL: https://codereview.chromium.org/1095983002

Cr-Commit-Position: refs/heads/master@{#326242}